// hr-frontend/src/lib/api/generated/model/jobPositionDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { JobPositionAssignmentView } from './jobPositionAssignmentView';

export interface JobPositionDetail {
  positionId?: number;
  code?: string;
  titleEn?: string;
  titleAr?: string;
  organizationId?: number;
  organizationCode?: string;
  organizationName?: string;
  orgUnitId?: number;
  orgUnitCode?: string;
  orgUnitName?: string;
  orgUnitNameAr?: string;
  gradeId?: number;
  gradeCode?: string;
  gradeName?: string;
  gradeNameAr?: string;
  gradeRank?: number;
  positionLevel?: number;
  reportsToPositionId?: number;
  reportsToPositionCode?: string;
  reportsToPositionTitle?: string;
  descriptionEn?: string;
  descriptionAr?: string;
  open?: boolean;
  active?: boolean;
  occupancyStatus?: string;
  currentAssignment?: JobPositionAssignmentView;
  directReportsCount?: number;
  assignmentHistoryCount?: number;
  createdAt?: string;
  updatedAt?: string;
}