// hr-frontend/src/lib/api/generated/model/jobPositionHierarchyNode.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface JobPositionHierarchyNode {
  positionId?: number;
  code?: string;
  titleEn?: string;
  titleAr?: string;
  positionLevel?: number;
  orgUnitId?: number;
  orgUnitName?: string;
  gradeId?: number;
  gradeCode?: string;
  gradeName?: string;
  assignedEmployeeId?: number;
  assignedEmployeeNumber?: string;
  assignedEmployeeName?: string;
  open?: boolean;
  active?: boolean;
  children?: JobPositionHierarchyNode[];
}